+++ Realm/RLMObjectStore.mm
@@ -25,7 +25,7 @@
 #import "RLMQueryUtil.hpp"
 #import "RLMUtil.hpp"
 
-#import <objc/runtime.h>
+#import <objc/message.h>
 
 static void RLMVerifyAndAlignColumns(RLMObjectSchema *tableSchema, RLMObjectSchema *objectSchema) {
     NSMutableArray *properties = [NSMutableArray arrayWithCapacity:objectSchema.properties.count];
@@ -58,9 +58,10 @@ static void RLMVerifyAndAlignColumns(RLMObjectSchema *tableSchema, RLMObjectSche
             }
         }
 
-        // align
-        schemaProp.column = tableProp.column;
-        [properties addObject:schemaProp];
+        // create new property with aligned column
+        RLMProperty *prop = [schemaProp copy];
+        prop.column = tableProp.column;
+        [properties addObject:prop];
     }
 
     // check for new missing properties
@@ -77,7 +78,8 @@ static void RLMVerifyAndAlignColumns(RLMObjectSchema *tableSchema, RLMObjectSche
                                                objectSchema.className, [exceptionMessages componentsJoinedByString:@"\n- "]]
                                      userInfo:nil];
     }
-    // ensure the order of the properties matches the column order in the table
+
+    // set new properties array with correct column alignment
     objectSchema.properties = properties;
 }
 
@@ -116,6 +118,7 @@ void RLMRealmCreateAccessors(RLMSchema *schema) {
     dispatch_once(&onceToken, ^{
         s_accessorSchema = [NSMutableArray new];
     });
+
     // create accessors for non-dynamic realms
     RLMSchema *matchingSchema = nil;
     for (RLMSchema *accessorSchema in s_accessorSchema) {
@@ -145,9 +148,10 @@ void RLMRealmSetSchema(RLMRealm *realm, RLMSchema *targetSchema, bool verify) {
     realm.schema = [targetSchema copy];
 
     for (RLMObjectSchema *objectSchema in realm.schema.objectSchema) {
+        objectSchema.realm = realm;
+
         // read-only realms may be missing tables entirely
-        objectSchema->_table = RLMTableForObjectClass(realm, objectSchema.className);
-        if (objectSchema->_table && verify) {
+        if (verify && objectSchema.table) {
             RLMObjectSchema *tableSchema = [RLMObjectSchema schemaFromTableForClassName:objectSchema.className realm:realm];
             RLMVerifyAndAlignColumns(tableSchema, objectSchema);
         }
@@ -159,12 +163,10 @@ void RLMRealmCreateTables(RLMRealm *realm, RLMSchema *targetSchema, bool updateE
     realm.schema = [targetSchema copy];
 
     // first pass to create missing tables
-    bool changed = false;
     NSMutableArray *objectSchemaToUpdate = [NSMutableArray array];
     for (RLMObjectSchema *objectSchema in realm.schema.objectSchema) {
         bool created = false;
         objectSchema->_table = RLMTableForObjectClass(realm, objectSchema.className, created);
-        changed |= created;
 
         // we will modify tables for any new objectSchema (table was created) or for all if updateExisting is true
         if (updateExisting || created) {
@@ -181,7 +183,6 @@ void RLMRealmCreateTables(RLMRealm *realm, RLMSchema *targetSchema, bool updateE
             // add any new properties (new name or different type)
             if (!tableSchema[prop.name] || ![prop isEqualToProperty:tableSchema[prop.name]]) {
                 RLMCreateColumn(realm, *objectSchema->_table, prop);
-                changed = true;
             }
         }
 
@@ -190,7 +191,6 @@ void RLMRealmCreateTables(RLMRealm *realm, RLMSchema *targetSchema, bool updateE
             RLMProperty *prop = tableSchema.properties[i];
             if (!objectSchema[prop.name] || ![prop isEqualToProperty:objectSchema[prop.name]]) {
                 objectSchema->_table->remove_column(prop.column);
-                changed = true;
             }
         }
 
@@ -199,21 +199,18 @@ void RLMRealmCreateTables(RLMRealm *realm, RLMSchema *targetSchema, bool updateE
             // if there is a primary key set, check if it is the same as the old key
             if (tableSchema.primaryKeyProperty == nil || ![tableSchema.primaryKeyProperty isEqual:objectSchema.primaryKeyProperty]) {
                 RLMRealmSetPrimaryKeyForObjectClass(realm, objectSchema.className, objectSchema.primaryKeyProperty.name);
-                changed = true;
             }
         }
         else if (tableSchema.primaryKeyProperty) {
             // there is no primary key, so if thre was one nil out
             RLMRealmSetPrimaryKeyForObjectClass(realm, objectSchema.objectClass, nil);
-            changed = true;
         }
     }
 
     // FIXME - remove deleted tables
-    for (RLMObjectSchema *objectSchema in realm.schema.objectSchema) {
-        // cache table instances on objectSchema
-        objectSchema->_table = RLMTableForObjectClass(realm, objectSchema.className);
 
+    for (RLMObjectSchema *objectSchema in realm.schema.objectSchema) {
+        objectSchema.realm = realm;
         RLMObjectSchema *tableSchema = [RLMObjectSchema schemaFromTableForClassName:objectSchema.className realm:realm];
         RLMVerifyAndAlignColumns(tableSchema, objectSchema);
     }
@@ -234,7 +231,7 @@ template<typename F>
 static inline NSUInteger RLMCreateOrGetRowForObject(RLMObjectSchema *schema, F primaryValueGetter, RLMSetFlag options, bool &created) {
     // try to get existing row if updating
     size_t rowIndex = tightdb::not_found;
-    tightdb::Table &table = *schema->_table;
+    tightdb::Table &table = *schema.table;
     RLMProperty *primaryProperty = schema.primaryKeyProperty;
     if ((options & RLMSetFlagUpdateOrCreate) && primaryProperty) {
         // get primary value
@@ -264,9 +261,9 @@ void RLMAddObjectToRealm(RLMObject *object, RLMRealm *realm, RLMSetFlag options)
     RLMVerifyInWriteTransaction(realm);
 
     // verify that object is standalone
-    if (object.deletedFromRealm) {
+    if (object.invalidated) {
         @throw [NSException exceptionWithName:@"RLMException"
-                                       reason:@"Adding a deleted object to a Realm is not permitted"
+                                       reason:@"Adding a deleted or invalidated object to a Realm is not permitted"
                                      userInfo:nil];
     }
     if (object.realm) {
@@ -289,7 +286,7 @@ void RLMAddObjectToRealm(RLMObject *object, RLMRealm *realm, RLMSetFlag options)
     // get or create row
     bool created;
     auto primaryGetter = [=](RLMProperty *p) { return [object valueForKey:p.getterName]; };
-    object->_row = (*schema->_table)[RLMCreateOrGetRowForObject(schema, primaryGetter, options, created)];
+    object->_row = (*schema.table)[RLMCreateOrGetRowForObject(schema, primaryGetter, options, created)];
 
     // populate all properties
     for (RLMProperty *prop in schema.properties) {
@@ -311,6 +308,15 @@ void RLMAddObjectToRealm(RLMObject *object, RLMRealm *realm, RLMSetFlag options)
         // skip primary key when updating since it doesn't change
         if (created || !prop.isPrimary) {
             RLMDynamicSet(object, prop, value, options | (prop.isPrimary ? RLMSetFlagEnforceUnique : 0));
+        }
+
+        // set the ivars for object and array properties to nil as otherwise the
+        // accessors retain objects that are no longer accessible via the properties
+        // this is mainly an issue when the object graph being added has cycles,
+        // as it's not obvious that the user has to set the *ivars* to nil to
+        // avoid leaking memory
+        if (prop.type == RLMPropertyTypeObject || prop.type == RLMPropertyTypeArray) {
+            ((void(*)(id, SEL, id))objc_msgSend)(object, prop.setterSel, nil);
         }
     }
 
@@ -335,7 +341,7 @@ RLMObject *RLMCreateObjectInRealmWithValue(RLMRealm *realm, NSString *className,
         // get or create our accessor
         bool created;
         auto primaryGetter = [=](RLMProperty *p) { return array[p.column]; };
-        object->_row = (*objectSchema->_table)[RLMCreateOrGetRowForObject(objectSchema, primaryGetter, options, created)];
+        object->_row = (*objectSchema.table)[RLMCreateOrGetRowForObject(objectSchema, primaryGetter, options, created)];
 
         // populate
         NSArray *props = objectSchema.properties;
@@ -352,7 +358,7 @@ RLMObject *RLMCreateObjectInRealmWithValue(RLMRealm *realm, NSString *className,
         // get or create our accessor
         bool created;
         auto primaryGetter = [=](RLMProperty *p) { return [value valueForKey:p.name]; };
-        object->_row = (*objectSchema->_table)[RLMCreateOrGetRowForObject(objectSchema, primaryGetter, options, created)];
+        object->_row = (*objectSchema.table)[RLMCreateOrGetRowForObject(objectSchema, primaryGetter, options, created)];
 
         // assume dictionary or object with kvc properties
         NSDictionary *dict = RLMValidatedDictionaryForObjectSchema(value, objectSchema, schema, !created);
@@ -395,7 +401,7 @@ void RLMDeleteAllObjectsFromRealm(RLMRealm *realm) {
 
     // clear table for each object schema
     for (RLMObjectSchema *objectSchema in realm.schema.objectSchema) {
-        objectSchema->_table->clear();
+        objectSchema.table->clear();
     }
 }
 
@@ -404,12 +410,12 @@ RLMResults *RLMGetObjects(RLMRealm *realm, NSString *objectClassName, NSPredicat
 
     // create view from table and predicate
     RLMObjectSchema *objectSchema = realm.schema[objectClassName];
-    if (!objectSchema->_table) {
+    if (!objectSchema.table) {
         // read-only realms may be missing tables since we can't add any
         // missing ones on init
         return [RLMEmptyResults emptyResultsWithObjectClassName:objectClassName realm:realm];
     }
-    tightdb::Query query = objectSchema->_table->where();
+    tightdb::Query query = objectSchema.table->where();
     RLMUpdateQueryWithPredicate(&query, predicate, realm.schema, objectSchema);
     
     // create and populate array
@@ -430,7 +436,7 @@ id RLMGetObject(RLMRealm *realm, NSString *objectClassName, id key) {
         @throw [NSException exceptionWithName:@"RLMException" reason:msg userInfo:nil];
     }
 
-    if (!objectSchema->_table) {
+    if (!objectSchema.table) {
         // read-only realms may be missing tables since we can't add any
         // missing ones on init
         return nil;
@@ -439,7 +445,7 @@ id RLMGetObject(RLMRealm *realm, NSString *objectClassName, id key) {
     size_t row = tightdb::not_found;
     if (primaryProperty.type == RLMPropertyTypeString) {
         if (NSString *str = RLMDynamicCast<NSString>(key)) {
-            row = objectSchema->_table->find_first_string(primaryProperty.column, RLMStringDataWithNSString(str));
+            row = objectSchema.table->find_first_string(primaryProperty.column, RLMStringDataWithNSString(str));
         }
         else {
             @throw [NSException exceptionWithName:@"RLMException"
@@ -449,7 +455,7 @@ id RLMGetObject(RLMRealm *realm, NSString *objectClassName, id key) {
     }
     else {
         if (NSNumber *number = RLMDynamicCast<NSNumber>(key)) {
-            row = objectSchema->_table->find_first_int(primaryProperty.column, number.longLongValue);
+            row = objectSchema.table->find_first_int(primaryProperty.column, number.longLongValue);
         }
         else {
             @throw [NSException exceptionWithName:@"RLMException"
@@ -477,6 +483,6 @@ RLMObject *RLMCreateObjectAccessor(__unsafe_unretained RLMRealm *realm,
                                    __unsafe_unretained RLMObjectSchema *objectSchema,
                                    NSUInteger index) {
     RLMObject *accessor = [[objectSchema.accessorClass alloc] initWithRealm:realm schema:objectSchema defaultValues:NO];
-    accessor->_row = (*objectSchema->_table)[index];
+    accessor->_row = (*objectSchema.table)[index];
     return accessor;
 }