+++ tools/RealmBrowser/RealmBrowserTests/RLMTestObjects.h
@@ -16,15 +16,9 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-#import "RLMTypeNode.h"
+#import <Realm/Realm.h>
 
-@class RLMArrayNode;
-@interface RLMObjectNode : RLMTypeNode
-
-@property (nonatomic) id<RLMRealmOutlineNode> childNode;
-@property (nonatomic) id<RLMRealmOutlineNode> parentNode;
-
-- (instancetype)initWithObject:(RLMObject *)object realm:(RLMRealm *)realm;
-- (RLMArrayNode *)displayChildArrayFromProperty:(RLMProperty *)property object:(RLMObject *)object;
-
+@interface RealmObject1 : RLMObject
+@property NSInteger intProp;
+@property NSString *stringProp;
 @end