+++ tools/RealmBrowser/RealmBrowserTests/RLMTestObjects.m
@@ -16,15 +16,7 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-#import "RLMTypeNode.h"
+#import "RLMTestObjects.h"
 
-@class RLMArrayNode;
-@interface RLMObjectNode : RLMTypeNode
-
-@property (nonatomic) id<RLMRealmOutlineNode> childNode;
-@property (nonatomic) id<RLMRealmOutlineNode> parentNode;
-
-- (instancetype)initWithObject:(RLMObject *)object realm:(RLMRealm *)realm;
-- (RLMArrayNode *)displayChildArrayFromProperty:(RLMProperty *)property object:(RLMObject *)object;
-
+@implementation RealmObject1
 @end